+++ grafana-plugin/src/containers/IntegrationForm/IntegrationForm.tsx
@@ -109,8 +109,6 @@ export const IntegrationForm = observer((props: IntegrationFormProps) => {
     extraGFormProps.customFieldSectionRenderer = CustomFieldSectionRenderer;
   }
 
-  console.log('Form DISPLAY []');
-
   return (
     <>
       {showIntegrarionsListDrawer && (
@@ -140,9 +138,7 @@ export const IntegrationForm = observer((props: IntegrationFormProps) => {
         <Drawer scrollableContent title={getTitle()} onClose={onHide} closeOnMaskClick={false} width="640px">
           <div className={cx('content')}>
             <VerticalGroup>
-              <HookForm navigateToAlertGroupLabels={navigateToAlertGroupLabels} />
-
-              {/* {isTableView && <HowTheIntegrationWorks selectedOption={selectedOption} />} */}
+              <HookForm navigateToAlertGroupLabels={navigateToAlertGroupLabels} selectedOption={selectedOption} />
             </VerticalGroup>
           </div>
         </Drawer>
@@ -375,49 +371,6 @@ const CustomFieldSectionRenderer: React.FC<CustomFieldSectionRendererProps> = ({
     setState({ selectedContactPointOption: option.value });
     setValue('contact_point', option.value);
   }
-};
-
-const HowTheIntegrationWorks: React.FC<{ selectedOption: ApiSchemas['AlertReceiveChannelIntegrationOptions'] }> = ({
-  selectedOption,
-}) => {
-  if (!selectedOption) {
-    return null;
-  }
-
-  return (
-    <Collapse
-      headerWithBackground
-      className={cx('collapse')}
-      isOpen={false}
-      label={<Text type="link">How the integration works</Text>}
-      contentClassName={cx('collapsable-content')}
-    >
-      <Text type="secondary">
-        The integration will generate the following:
-        <ul className={cx('integration-info-list')}>
-          <li className={cx('integration-info-item')}>Unique URL endpoint for receiving alerts </li>
-          <li className={cx('integration-info-item')}>
-            Templates to interpret alerts, tailored for {selectedOption.display_name}{' '}
-          </li>
-          <li className={cx('integration-info-item')}>{selectedOption.display_name} contact point </li>
-          <li className={cx('integration-info-item')}>{selectedOption.display_name} notification</li>
-        </ul>
-        What you'll need to do next:
-        <ul className={cx('integration-info-list')}>
-          <li className={cx('integration-info-item')}>
-            Finish connecting Monitoring system using Unique URL that will be provided on the next step{' '}
-          </li>
-          <li className={cx('integration-info-item')}>
-            Set up routes that are based on alert content, such as severity, region, and service{' '}
-          </li>
-          <li className={cx('integration-info-item')}>Connect escalation chains to the routes</li>
-          <li className={cx('integration-info-item')}>
-            Review templates and personalize according to your requirements
-          </li>
-        </ul>
-      </Text>
-    </Collapse>
-  );
 };
 
 const IntegrationBlocks: React.FC<{
